+++ src/OneDriveShortcuts.psm1
@@ -1,5 +1,5 @@
-#Requires -Version 5.1
-#Requires -Modules MSAL.PS
+#Requires -Version 7.4
+#Requires -Modules PSMSALNet
 $Public = @(Get-ChildItem -Path (Join-Path -Path $PSScriptRoot -ChildPath "\public\*.ps1") -ErrorAction SilentlyContinue)
 $Private = @(Get-ChildItem -Path (Join-Path -Path $PSScriptRoot -ChildPath "\private\*.ps1") -ErrorAction SilentlyContinue)